feat: tauri command get_file_icon()
